backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
torch==2.6.0
torchvision
pydantic[email]>=2.11.7
pydantic-settings>=2.0.0
supabase==2.28.0
python-dotenv==1.2.1
bcrypt
python-multipart
pytest
pytest-asyncio
